No reason to throw up an IOError if the old file doesn't exist anymore.  Should we be deleting these items at startup?
git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1663 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 58ed1335112cc9b83797696fce68ababbee23c9e

+++ tv/portable/downloader.py
@@ -205,16 +205,17 @@ URL was %s""" % self.url
 WARNING: can't migrate download because we don't have a filename!
 URL was %s""" % self.url
                 return
-            newfilename = os.path.join(config.get(prefs.MOVIES_DIRECTORY),
-                    shortFilename)
-            newfilename = nextFreeFilename(newfilename)
-            try:
-                shutil.move(filename, newfilename)
-            except IOError, error:
-                print "WARNING: Error moving %s to %s (%s)" % (self.status['filename'],
-                        newfilename, error)
-            else:
-                self.status['filename'] = newfilename
+            if os.path.exists(filename):
+                newfilename = os.path.join(config.get(prefs.MOVIES_DIRECTORY),
+                        shortFilename)
+                newfilename = nextFreeFilename(newfilename)
+                try:
+                    shutil.move(filename, newfilename)
+                except IOError, error:
+                    print "WARNING: Error moving %s to %s (%s)" % (self.status['filename'],
+                            newfilename, error)
+                else:
+                    self.status['filename'] = newfilename
 
     ##
     # Removes downloader from the database
